backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Darf Vermieter Das? Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "google-generativeai>=0.8.4",
    "openai>=1.61.0",
    "pillow>=11.1.0",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
]
